utils/sigma-converter/requirements.txt
pySigma>=0.11,<1.0
sigma-cli>=1.0,<2.0
pySigma-backend-splunk>=1.0,<2.0
pySigma-backend-opensearch>=1.0,<2.0
pySigma-backend-microsoft365defender>=0.3,<1.0
pyyaml>=6.0
pytest>=7.0
